Expose item construction helper in the Player package
