+++ Package.swift
@@ -66,7 +66,8 @@ let package = Package(
             name: "CoreBusiness",
             dependencies: [
                 .target(name: "Analytics"),
-                .target(name: "Diagnostics")
+                .target(name: "Diagnostics"),
+                .target(name: "Player")
             ]
         ),
         .target(
@@ -123,8 +124,7 @@ let package = Package(
             name: "CoreBusinessTests",
             dependencies: [
                 .target(name: "Circumspect"),
-                .target(name: "CoreBusiness"),
-                .target(name: "Player")
+                .target(name: "CoreBusiness")
             ]
         ),
         .testTarget(
